+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - Disable Biconomy for CERE ERC20 transfer
 - Application bundle optimization to improve load speed
+- Add CERE ERC20 token to default assets
+- Disable OPT code auto-capitalization
 
 ### v1.21.0