+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Application bundle optimization to improve load speed
 - Add CERE ERC20 token to default assets
 - Disable OPT code auto-capitalization
+- Optimize login flow to decrease time to login
 
 ### v1.21.0