+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add CERE ERC20 token to default assets
 - Disable OPT code auto-capitalization
 - Optimize login flow to decrease time to login
+- Increase assets balance pooling interval to 10 seconds
 
 ### v1.21.0